Return double value for sensor distance

+++ sharpSensorCommWMedian/SharpIRNormal/SharpIR.h
@@ -27,7 +27,7 @@ class SharpIR
   public:
 
     SharpIR (int irPin, long sensorModel);
-    int distance();
+    double distance();
 
   private:
